requirements.txt
fastapi
uvicorn
openai
faiss-cpu
pypdf2
python-docx
extract-msg
